Fixed yellow builds (int->long)
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@5646 a1c6a512-1295-4272-9138-f99709370657

+++ apps/main_menu.c
@@ -144,7 +144,7 @@ bool show_info(void)
     bool done = false;
     int key;
     int state = 1;
-    unsigned int size, free;
+    unsigned long size, free;
 
     fat_size( IF_MV2(0,) &size, &free );
     size /= 1024;
